+++ com.unity.shadergraph/Editor/Data/Graphs/GraphData.cs
@@ -1426,6 +1426,7 @@ namespace UnityEditor.ShaderGraph
                 copyProp.precision = sourceProp.precision;
                 copyProp.overrideHLSLDeclaration = sourceProp.overrideHLSLDeclaration;
                 copyProp.hlslDeclarationOverride = sourceProp.hlslDeclarationOverride;
+                copyProp.useCustomSlotLabel = sourceProp.useCustomSlotLabel;
             }
 
             // sanitize the display name (we let the .Copy() function actually copy the display name over)